Tested, probably OK: pkg::upgrade() | print::changes() | wallet functions

+++ src/pkg/remove.sh
@@ -30,37 +30,52 @@ pkg::remove::prompt() {
 
 	# CREATE REMOVE LIST
 	local REMOVE_LIST || return 1
-	if [[ $OPTION_REMOVE_BASH = true ]]; then
+	if [[ $OPTION_BASH = true ]]; then
 		printf "${OFF}%s\n%s${BYELLOW}%s${OFF}%s\n" \
 			"[monero-bash] ($MONERO_BASH_VER) cannot be removed normally" \
-			"type: " \
-			"[monero-bash uninstall]" \
+			"Type: " \
+			"[monero-bash uninstall] " \
 			"to cleanly uninstall monero-bash from your system"
 		exit 1
 	fi
-	[[ $OPTION_REMOVE_MONERO = true ]] && REMOVE_LIST="$REMOVE_LIST [monero]"
-	[[ $OPTION_REMOVE_P2POOL = true ]] && REMOVE_LIST="$REMOVE_LIST [p2pool]"
-	[[ $OPTION_REMOVE_XMRIG = true ]]  && REMOVE_LIST="$REMOVE_LIST [xmrig]"
+	[[ $OPTION_MONERO = true ]] && REMOVE_LIST="$REMOVE_LIST monero"
+	[[ $OPTION_P2POOL = true ]] && REMOVE_LIST="$REMOVE_LIST p2pool"
+	[[ $OPTION_XMRIG = true ]]  && REMOVE_LIST="$REMOVE_LIST xmrig"
 
 	# CHECK IF PACKAGE IS INSTALLED
-	if [[ $OPTION_REMOVE_MONERO = true && -z $MONERO_VER ]]; then
-		printf "${OFF}%s\n" "[monero] is not installed"
-		REMOVE_LIST="${REMOVE_LIST/\[monero\]}"
+	if [[ $OPTION_MONERO = true && -z $MONERO_VER ]]; then
+		printf "${OFF}%s\n" "[Monero] is not installed"
+		REMOVE_LIST="${REMOVE_LIST//monero}"
 	fi
-	if [[ $OPTION_REMOVE_P2POOL = true && -z $P2POOL_VER ]]; then
-		printf "${OFF}%s\n" "[p2pool] is not installed"
-		REMOVE_LIST="${REMOVE_LIST/\[p2pool\]}"
+	if [[ $OPTION_P2POOL = true && -z $P2POOL_VER ]]; then
+		printf "${OFF}%s\n" "[P2Pool] is not installed"
+		REMOVE_LIST="${REMOVE_LIST//p2pool}"
 	fi
-	if [[ $OPTION_REMOVE_XMRIG = true && -z $XMRIG_VER ]]; then
-		printf "${OFF}%s\n" "[xmrig] is not installed"
-		REMOVE_LIST="${REMOVE_LIST/\[xmrig\]}"
+	if [[ $OPTION_XMRIG = true && -z $XMRIG_VER ]]; then
+		printf "${OFF}%s\n" "[XMRig] is not installed"
+		REMOVE_LIST="${REMOVE_LIST//xmrig}"
 	fi
-	[[ $REMOVE_LIST = " " ]] && exit 1
+
+	# EXIT ON EMPTY LIST
+	if [[ $REMOVE_LIST =~ ^[[:space:]]+$ || -z $REMOVE_LIST ]]; then
+		log::debug "REMOVE_LIST is empty, exiting"
+		exit 1
+	fi
+
+	# CREATE UI LIST
+	local PROMPT_REMOVE_LIST i
+	for i in $REMOVE_LIST; do
+		case "$i" in
+			monero) PROMPT_REMOVE_LIST="$PROMPT_REMOVE_LIST  [Monero]";;
+			p2pool) PROMPT_REMOVE_LIST="$PROMPT_REMOVE_LIST  [P2Pool]";;
+			xmrig) PROMPT_REMOVE_LIST="$PROMPT_REMOVE_LIST  [XMRig]";;
+		esac
+	done
 
 	# PROMPT
 	printf "${BWHITE}%s${OFF}%s\n\n${BWHITE}%s" \
 		"Packages to remove: " \
-		"$REMOVE_LIST" \
+		"$PROMPT_REMOVE_LIST" \
 		"Continue with removal? (y/N) "
 	if ask::no; then
 		print::exit "Canceling installation"
@@ -72,41 +87,43 @@ pkg::remove::prompt() {
 	fi
 
 	# START REMOVAL
-	[[ $OPTION_REMOVE_MONERO = true ]] && struct::pkg monero && pkg::remove
-	[[ $OPTION_REMOVE_P2POOL = true ]] && struct::pkg p2pool && pkg::remove
-	[[ $OPTION_REMOVE_XMRIG = true ]]  && struct::pkg xmrig  && pkg::remove
+	for i in $REMOVE_LIST; do
+		struct::pkg $i
+		pkg::remove
+	done
+	exit 0
 }
 
 # pkg::remove packages
 # called from pkg::remove::prompt()
 pkg::remove() {
-	log::debug "starting ${FUNCNAME}() on package: ${PKG[name]}"
+	log::debug "${PKG[pretty]} | starting"
 	trap "" INT
-	print::remove
+	print::pkg::remove
 
 	# SET TRAP
-	trap 'trap::remove &' EXIT
+	trap 'trap::remove' EXIT
 
 	# REMOVE PKG DIRECTORY
-	log::prog "${PKG[directory]}..."
-	rm "${PKG[directory]}"
-	log::ok "${PKG[directory]} deleted"
+	log::prog "Removing ${PKG[directory]}..."
+	rm -rf "${PKG[directory]}"
+	log::ok "Removed ${PKG[directory]}"
 
 	# REMOVE SYSTEMD SERVICE
-	log::prog "${PKG[service]}..."
-	sudo rm "$SYSTEMD/${PKG[service]}"
-	log::ok "${PKG[service]} deleted"
+	log::prog "Removing ${PKG[service]}..."
+	sudo rm -f "${SYSTEMD:?}/${PKG[service]:?}"
+	log::ok "Removed ${PKG[service]}"
 
 	# REMOVE CHANGELOG
-	log::prog "${PKG[pretty]} changelog..."
-	rm "$CHANGES/${PKG[name]}"
-	log::ok
+	log::prog "Removing ${PKG[pretty]} changelog..."
+	rm -rf "${CHANGES:?}/${PKG[name]:?}"
+	log::ok "Removed ${PKG[pretty]} changelog"
 
 	# UPDATE LOCAL STATE
 	log::prog "Updating local state..."
 	sudo sed \
-		-i -e "s/${PKG[var]}_VER=./${PKG[var]}_VER=/" "$STATE" \
-		-i -e "s/${PKG[var]}_OLD=./${PKG[var]}_OLD=\"true\"/" "$STATE"
+		-i -e "s/${PKG[var]}_VER=.*/${PKG[var]}_VER=/" "$STATE" \
+		-i -e "s/${PKG[var]}_OLD=.*/${PKG[var]}_OLD=\"true\"/" "$STATE"
 	log::ok "Updated local state"
 
 	# FREE TRAP
@@ -118,6 +135,6 @@ pkg::remove() {
 	log::ok "Reloaded systemd"
 
 	# END
-	print::removed
-	exit 0
+	print::pkg::removed
+	return 0
 }
